Refactor into query, execute, shared rs files
