+++ contracts/admin/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod contract;
+pub mod execute;
+pub mod query;
+pub mod shared;
